Separate App and Server operations
